Replace single line .PHONY
The old imlementation used .PHONY as a one line.
This approach cases problem since the line keeps growing
and the list gets not updated. As example it was missing
the target export-cross and contained not existing targets
like boot and auto_config. Separating .PHONY and add it
above every phony target itself will prevent this from
happing