Adding a test if CROSS_PREFIX is writeable
This commit adds a test to prepare() for checking
if the given CROSS_PREFIX is writeable. The test
must be executed before build_target() to ensure
that the build process has not started yet. This
commit will also move the default path of CROSS_PREFIX
to the beginning of the file for prevent code
duplication. Namely in the updated prepare(),
build_target() and test_version()