Adds the make target toolchain-build and toolchain-sanity
The target toolchain-build is a shortcut for tools/toolchain.sh.
This will make sure that the script is not executed by the user
and therefore the relative paths are correct (as example downloads
and PKG).

The target toolchain-sanity will test if the installed version
of the toolchain is installed and up-to-date. This target is
a dependency from 'common' which will ensure that the sanity
test is always executed before compilation.